Fix "fedpkg help" command (make it work again) (#681242)
This changes the parser_help command to a lambda which passes
the parser variable to the usage() function as a second
parameter.

We forgot to give the usage() function its "parser" parameter,
and after we moved the definition of the parser variable out
of the main module, the usage() function did not see the parser
variable any more.