Fix mockconfig property
Some plugins (namely fedpkg) try to set the Commands.mockconfig
attribute, which fails now that it became a property and that we moved
to new-style classes.

The solution is either to add a property setter in rpkg, or to modify
all those clients so they do it differently (e.g by overloading
load_mockconfig so it accepts an optional config argument, and use that
instead).

The former is less invasive and allows keeping compatibility for
existing plugins.