Ability to skip NVR construction altogether for builds
If one doesn't need to construct NVR for build it shouldn't stop us to
submit this to build system.

Let's say we use some macros which are not defined on our local host:

%global foo foo_package_name
%foo_package %foo

Until now rpm would complain:

error: line 2: Unknown tag: %foo_package foo_package_name

and we won't be able to continue with build. But this macro is defined
in some package in builder's buildroot so it should be possible to build
this package.

When developer use --skip-nvr-check option we don't care about errors in
NVR construction and continue. If package is build from SRPM we will
refer to this build by SRPM basename, otherwise we will use package
name.