gitignore: Make sure each line ends with a \n
When loading an existing file, if its last line didn't end with a \n
character, then we'd load it as it is.

Then, if we added a new line, it would be appended to the last existing
line, as one line, which is clearly not what we want.

So we must ensure every line ends with a \n character when we load an
existing file.

While we're at it, if we're going to check for \n in 2 places, we might
as well make that a new function.