tests: Ensure functioning of Commands.clone
A lot of stuff to figure out (TODO) in there, but this actually manages
to test the function quite nicely.

When adding more tests for other functions, some of this stuff will
eventually get factored out and figured out better, but for now... this
works and covers almost 100% of the function. :)