tests: Ensure proper functioning of Commands.add_tag
The tests fail!

That's because they just found 1 bug in the code. :)

However, notice the game I'm playing with the EDITOR environment
variable. This is a gross hack, but it is made necessary by a bad
design decision in the pyrpkg API: a library should never do some
interactive stuff in subprocesses like that.