add_tag: Run the tag command in the right directory
We were running `git tag` in the current directory.

This doesn't cause any problem when we run the tool in the git directory
we want to tag, which is the most common case.

However, when using the API (for example in the unit tests), the current
directory might not be the git clone, we might really be anywhere.

As such, the `git tag` command would just fail.

Worse, when running the tests, we would add tags to the clone of
upstream rpkg!