list_tags: Stop executing a command
We have a Git API, let's use it, instead of constantly running commands
in subprocesses.

This even results in simpler code!

In addition, it makes the method actually testable. Without this, the
tag list was printed in the stdout **of the subprocess**, which is
unaccessible to the parent process, and as such to the tests. The tests
couldn't even verify what the library had done!