tests: Ensure functioning of Commands.list_tag
This is pretty horrible though. The library function is just a poorly
thought-out API. A proper API would return the results, and let the
caller handle them (for example, printing them to stdout, or in a web
interface, or...).

However, the rpkg API instead directly prints it to stdout, which not
only prevents consumers from easily doing something nice with it, but
also prevents **unit tests** (!!) from easily verifying the behaviour of
the function.

As a result, if we want unit tests, we need to do something pretty
horrible: hijack the stdout of the function, and parse it.