sources: Move to the new file format
We are eventually going to move away from md5 for the sources. However,
in order to make the migration (and future ones) easier, we want to
indicate on each line of the 'sources' file what is the hash function
used to compute the hash of the file.

Fortunately, the md5sum/sha512sum/... utilities support two file formats
as their inputs and outputs:

  * the current format:
    `ahash  afile`

  * the BSD-style format, obtained with the `--tag` option:
    `HASHTYPE (afile) = ahash

This second format is perfect for us, so this commit moves our 'sources'
file handling to it.

A couple of notes:

* we preserve compatibility with existing files, so lines in the old
  format are still read
* we now only ever write lines in the new format, which means that when
  reading an existing file, all currently present lines will be converted to
  the new format when writing back to the file