sources: Forbid mixing hash types
There theoretically isn't any problem with mixing various hashes, as far
as rpkg is concerned.

However, taking for example the following file:

    $ cat sources
    MD5 (pygit2-0.22.0.tar.gz) = 31dc65b3cbe6e39d75a39db86dd7cd8c
    SHA512 (pygit2-0.22.0.tar.gz) =2079adc3285bf08fc7227b412b9d3dc[...]

When someone tries verifying it with `md5sum -c` they would get:

    $ md5sum -c sources
    pygit2-0.22.0.tar.gz: OK
    md5sum: WARNING: 1 line is improperly formatted

They'd need to also to check the file with `sha512sum -c`, but then they
would get:

    $ sha512sum -c sources
    pygit2-0.22.0.tar.gz: OK
    sha512sum: WARNING: 1 line is improperly formatted

This is clearly not ideal, and it might cause confusion.

Mixing might happen when a file contains lines with the old hash, then
after the migration a packager tries to upload an additional source
file.

This commit prevents mixing, by forcing packagers to use `new-sources`
and redo the source file entirely with the new hash.

It's a one time annoyance, but in addition to reducing future confusion,
it helps moving towards the goal of completely dropping MD5.