tests: Properly open/close the file
In addition to being just cleaner, this removes the only warning when
running the tests for the pyrpkg.sources module with Python3.

Not that we're anywhere ready to have pyrpkg run with Python3, but any
little step in that direction is a good thing.

Plus it means pyrpkg.sources works fine with both Python2 and Python3!