Add a new lookaside module
At the moment this doesn't do much: it only stores the 3
lookaside-related parameters we have (a download url, an upload url and
a hash type), but eventually it will become a full-fledged lookaside
cache client.

Using a property for it makes it trivial for downstreams to override it
if they need to use their own implementation of a lookaside cache
client.

And using a cached property means we won't reinitialize the
CGILookasideCache object every time we access the
Commands.lookasidecache attribute.