lookaside: Move handling of file verification
The only time we need to verify whether a source file is valid is when
we interact with the lookaside cache, so it makes sense to have it
there, and remove some clutter from the main module.

However, the Commands._verify_file might be used by downstreams, so
let's keep it for a while, make it use the new lookaside module, and
emit a deprecation warning.