lookaside: Move handling of file hashing
The only time we need to compute the hash of a source file is when we
interact with the lookaside cache, so it makes sense to have it there,
and remove some clutter from the main module.

However, the Commands._hash_file might be used by downstream, so let's
keep it for a while, make it use the new lookaside module, and emit a
deprecation warning.