lookaside: Be more flexible when building the download URL
We're never going to know in advance everything our downstreams are
going to need as part of their download URL.

However, we already know that one of them, CentOS, needs the git branch
in there.

We could add the git branch as an argument to the download function, and
use it for the string interpolation.

However, someone would soon come and need something else.

This commit provides a much more generic way, which could work for all
downstreams. All they'd have to do is override pyrpkg.Commands.sources
to pass their own keyword arguments to
pyrpkg.lookaside.CGILookasideCache.download.

We could even add them to pyrpkg if they make sense to have them shared.

And in fact, since we already know that CentOS wants the git branch,
this commit also adds it, which should make them much happier, as that
should make it much easier for them to reuse more code from pyrpkg.