lookaside: Allow client-side and custom CA certificates
Some downstreams, most notably fedpkg, interact with a lookaside cache
which uses a self-signed certificate.

Some downstreams, still fedpkg, require an authentication for uploading
source files to the lookaside cache, based on a client-side certificate.

This commit makes it easier on this downstreams to reuse our
CGILookasideCache implementation.

All they need to do is define a cert_file and a ca_cert properties on
their pyrpkg.Commands subclass, and things are going to work
automatically.

And in fact, this is already what fedpkg does, so we're not breaking
anything.