lookaside: Check if a file already was uploaded
This is one more chunk of code which gets moved from pyrpkg.Commands to
pyrpkg.lookaside.CGILookasideCache, where it gets entirely tested.

It also gains some better error handling, as well as the capacity to use
a client-side certificate and/or a custom CA certificate: none are used
by default in pyrpkg, but a client like fedpkg automatically benefits
from this, without needing to do anything on its own.

The Commands._create_curl method is now completely unused, and is marked
as deprecated. However its implementation code is maintained, in case
downstreams were actually using it.