Stop making source files read-only
These lines are trying to ensure that we can read the file content, and
being able to read it is of course necessary if we're uploading it.

However, a few lines above, we actually computed the hash of the file,
which... required being able to read it.

So if the file wasn't readable, we'd have failed much earlier anyway,
which makes those two lines useless.

In addition, they make the file read-only, which is annoying for package
maintainers when try to remove it, as they then require a confirmation.