lookaside: Take over file uploads
This is the last piece of code related to the lookaside cache
interaction which needed to be moved.

In addition to just moving the code, we now better handle errors, we
exclusively use pycurl instead of calling the curl command, and this
whole thing is completely covered by unit tests.

It also automatically handles usage of client-side and CA certificates,
to make it easier for downstreams like fedpkg.

The Commands._do_curl method is now completely unused, and is marked as
deprecated. However its implementation code is maintained, in case
downstreams were actually using it.