gitignore: Properly handle adding matching lines
The GitIgnore.add method was trying to avoid adding lines which already
are in the file.

However, it wasn't trying to match wildcards, using the GitIgnore.match
method.

In addition, the GitIgnore user (Commands.upload) was trying to
GitIgnore.match before trying to add the line.

This commit makes GitIgnore.add check for matching lines properly, and
simplifies Commands.upload in consequence.