fix broken when non-ASCII in path
Path can be got from current directory or the --path option. In both
cases, unicode string is used instead of byte string.

This change is compatible with Python 3 by defining pyrpkg own's u and
getcwd methods, that is because it's necessary to convert path specified
via --path option from byte string to Unicode string in UTF-8, and get
current directory as Unicode string. So far, six.moves.getcwd is only
supported in 1.10.0 which is available for fc24, thus, six.moves.getcwd
is not available in any old distros using older version of six.

This is related to the fix of bug 1286374.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>