Simplify _run_command
Either TTY or non-TTY, _run_command does same thing that run a single
command or pipe command. This change simplifies _run_command to run
commands for both TTY or non-TTY. As a result, no output to stdout and
stderr is captured, user is able to redirect output to a destination by
himself.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>