Refine Kerberos with cccolutils
_has_krb_creds tells whether there is credentials in credential cache. The
result will be used to determine if Kerberos authentication could be used to
complete authentication against Koji. This patch simplifies the code to see the
existence of credentials, and both credential cache collection and simple file
cache are supported.

It is true as long as a credential exists in the cache, whatever it expires. If
it is, user will be notified and he needs to renew it manually.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>