Get target_arch in stable way
Use platform.machine to get target_arch from the machine where tests are
running. This is requires for running tests in various machines that
have different arches. It is useful for building packages in Copr and
Koji especially, that is because we have to build packages for i386,
e.g. fedora-24-i386, and the build might happen in an ARM builder
machine sometimes. So, we have to ensure to get

- target arch i686 on i386
- target arch armv7hl on ARM

Signed-off-by: Chenxiong Qi <cqi@redhat.com>