Ignore TestModulesCli if openidc-client is unavailable
This is for running tests in downstream client when openidc-client is
not used for OpenIDC authentication and not installed in system as well.
Especially the case of building rpkg for RHEL6/7 for rhpkg internal use.
In which case, Kerberos authentication is used instead, and no Python
openidc-client is installed to run tests while building RPM package.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>